feat(config): ip and memory allocation configurable via env vars
It is now possible to specify the private_network ip address for your vm
via an environment variable (IP). If no ip address is specified then a default
one will be applied (192.168.10.10).

The ammount of memory allocated to the vm is also configurable via an
environment variable (MEM). The default memory allocation is 2048.

The Vagrant configuration is now set up to sync any folder to the vm that is in
the parent directory.