feat(docker setup): allow insecure registry
Recent changes to Docker prevent pulling images from registries that are
not running over https. Docker classifies these registries as insecure.
If you are running a private registry without an ssl certificate then
this is obviously problematic.

In order to permit docker to pull from a non-secure private registry,
a provisioning recipe has been included that:

 * Reads a `REGISTRY_IP` environment variable, which if not specified
 will default to 192.168.10.10

 * Modifies the docker upstart conf file on the vm, to pass the
 `insecure-registry` for the registry IP, to the docker binary.

 * Issues a `fig pull --allow-insecure-ssl` command to ensure any
 "insecure" images specified in `/scripts/fig.yml` are downloaded.