chore(docker_setup): remove `docker-compose pull` instruction
Previously the `docker_setup` file concluded by calling `docker-compose
pull`. However now that we expect to be using a v2 docker registry over
ssl, we first need to login to that registry which requires manual entry
of user credentials. Therefore this login step is left as a manual step
to be done once the vm is provisioned.