fix(docker_setup): append registry host entry to `/etc/hosts`
Previously the action of adding the `private-registry.dev` host entry to
the `/etc/hosts` was unintentionally clobbering (overwrting) the
`/etc/hosts` file. This was a result of redirecting the output of the
`echo` command with the `>` operator. This has been fixed to append the
output to the `/etc/hosts` file through use of the `>>` operator.